backend/requirements.txt
fastapi
uvicorn
openai
playwright
python-dotenv
beautifulsoup4
pydantic
sqlalchemy
psycopg2-binary
